+++ src/providers/ftl-code-action-provider.ts
@@ -27,7 +27,27 @@ export class FtlCodeActionProvider implements CodeActionProvider {
       range: Range | Selection,
       context: CodeActionContext,
       token: CancellationToken): ProviderResult<(Command | CodeAction)[]> {
-    return this.changeListTypeCodeAction(document, range, context);
+    let result = this.changeListTypeCodeAction(document, range, context);
+    if (result) return result;
+    result = this.markAsUnused(document, range, context);
+    return result;
+  }
+
+  markAsUnused(
+      document: TextDocument,
+      range: Range | Selection,
+      context: CodeActionContext): CodeAction[] | undefined {
+    const unusedDefDiagnostic = context.diagnostics.find(diag => diag.code == FtlErrorCode.unusedRef);
+    if (!unusedDefDiagnostic) return;
+    const node = this.getNode(document, range);
+    if (!node) return;
+    const action = new CodeAction(`Disable unused reference warning`, CodeActionKind.QuickFix);
+    action.diagnostics = [unusedDefDiagnostic];
+    action.edit = new WorkspaceEdit();
+    action.edit.insert(document.uri,
+        document.positionAt((node.startTagEnd ?? node.end) - 1),
+        ` unused="true"`);
+    return [action];
   }
 
   changeListTypeCodeAction(
@@ -36,8 +56,8 @@ export class FtlCodeActionProvider implements CodeActionProvider {
       context: CodeActionContext): CodeAction[] | undefined {
     const invalidTypeDiagnostic = context.diagnostics.find((diag) => diag.code == FtlErrorCode.listTypeMismatch);
     if (!invalidTypeDiagnostic) return;
-    const htmlDocument = this.documentCache.getHtmlDocument(document);
-    let node: Node | undefined = htmlDocument.findNodeBefore(document.offsetAt(range.start));
+    let node = this.getNode(document, range);
+    if (!node) return;
     const findNode = transformModFindNode(node);
     if (findNode) node = findNode;
 
@@ -52,6 +72,11 @@ export class FtlCodeActionProvider implements CodeActionProvider {
         document.positionAt((node.startTagEnd ?? node.end) - 1),
         ` type="${BlueprintListTypeAny}"`);
     return [action];
+  }
+
+  getNode(document: TextDocument, range: Range | Selection): Node | undefined {
+    const htmlDocument = this.documentCache.getHtmlDocument(document);
+    return htmlDocument.findNodeBefore(document.offsetAt(range.start));
   }
 
   resolveCodeAction(codeAction: CodeAction, token: CancellationToken): ProviderResult<CodeAction> {